bug fixed added unsigned long

+++ 0x00-hello_world/6-size.c
@@ -6,10 +6,10 @@
 */
 int main(void)
 {
-printf("Size of char: %zu byte(s)\n", sizeof(char));
-printf("Size of int: %zu byte(s)\n", sizeof(int));
-printf("Size of long int: %u byte(s)\n", sizeof(long int));
-printf("Size of long long: %zu byte(s)\n", sizeof(long long));
-printf("Size of float: %zu byte(s)\n", sizeof(float));
+printf("Size of char: %lu byte(s)\n", (unsigned long)sizeof(char));
+printf("Size of int: %lu byte(s)\n", (unsigned long)sizeof(int));
+printf("Size of long int: %lu byte(s)\n", (unsigned long)sizeof(long int));
+printf("Size of long long: %lu byte(s)\n", (unsigned long)sizeof(long long));
+printf("Size of float: %lu byte(s)\n", (unsigned long)sizeof(float));
 return (0);
 }
